pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "merchwireanalytics"
version = "0.1.0"
description = "Shopify pricing and Meta ads analytics pipeline"
authors = [{name = "MerchWire", email = "team@merchwire.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.22.0",
    "httpx[http2]>=0.27.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.12",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.6",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "pandas>=2.1",
    "jinja2>=3.1",
    "python-dateutil>=2.8",
    "boto3>=1.34",
    "itsdangerous>=2.1",
    "typer>=0.9",
    "orjson>=3.9",
    "pendulum>=3.0"
]

[project.optional-dependencies]
email = ["mjml-python>=0.9"]
test = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "respx>=0.20",
]

[project.urls]
Homepage = "https://merchwire.com"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
]
asyncio_mode = "auto"

